refactor: extract migrateV1toBillV2 to billMigrations.ts
The fuzz tests were duplicating the migration logic, so a divergence
between the copy and the real function would go undetected. Moves
migrateV1toBillV2 into src/lib/billMigrations.ts and imports it from
both useBillSplit and the test suite.
